extract Tests-tab panels into Panels/ dir in example
